+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Version 1.0.2
-- Add a new optional flag to `test` command called `--xml`. It produces an xml output of the test results using the `xmlrunner` library.
+- Add a new optional flag to `test` command called `--xml`. It produces an xml output of the test results using the `unittest-xml-reporting` library .
 
 # Version 1.0.1
 - Bug fix in `WXFConsumer`.